bootloader: allow building without bootloader support
Some platforms we build snapd for have no use for boot or bootloader support,
since snapd is not playing any role in the management of those platform, nor it
participates in the boot process. This is a first step to enable building of a
slightly leaner version of snapd package.
